requirements.txt
Flask>=2.3.3
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.2
Flask-WTF>=1.1.1
WTForms>=3.0.1
SQLAlchemy>=2.0.20
python-dotenv>=1.0.0
email_validator>=2.1.0
scikit-learn>=1.3.2
pandas>=2.1.0
numpy>=1.24.3
Werkzeug>=2.3.7
plotly>=5.1.0
dash>=2.0.0
geopy>=2.4.0
gunicorn>=20.1.0
pytest>=6.2.5
folium>=0.14.0
requests>=2.31.0